update disassemblers + bench.png
capstone moves from 5.0-rc2 to 5.0.6
xed moves from v2023.07.09 to v2025.06.08

Python 3.12.3 complains about `\d` not being a valid escape sequence.
what it looks like is happening is that "\d" is interpreted as an
escaped `d` in the string and python warns that `\d` is not an escape
sequence. then Python does "what you mean" by parsing it as if it was written
`\\d`. *then* `re.match` gets `"\d"`, and this is why it had worked
quietly in the past.  now it warns and continues to do what we meant. to
avoid the problem and get the right string, prefix the regex with `r` so
no escaping is processed when parsing the string, and instead the
literal written regex makes it to `re.match`.

with these done and the recent `bddisasm` additions, i've rerun
benchmarks on my newer Linux machine and updated bench.png accordingly.
if we're gonna say #14 won't do for whatever reason, we should keep the
image up to date too :)